move ActantType + derivatives to @shared/enums
- updated client's webpack for aliasing the dir
- client files update to work with this enum/type
