+++ packages/client/src/pages/MainPage/containers/ActantSuggester/ActantSuggester.tsx
@@ -7,6 +7,7 @@ import { CActant } from "constructors";
 import { Entities } from "types";
 import { useQuery, useQueryClient } from "react-query";
 import api from "api";
+import { CategoryActantType } from "@shared/enums";
 
 interface ActantSuggesterI {
   categoryIds: string[];
@@ -82,7 +83,7 @@ export const ActantSuggester: React.FC<ActantSuggesterI> = ({
 
   const handleCreate = async (newCreated: {
     label: string;
-    category: "P" | "G" | "O" | "C" | "L" | "V" | "E";
+    category: CategoryActantType;
   }) => {
     const newActant = CActant(newCreated.category, newCreated.label);
     const resCreate = await api.actantsCreate(newActant);
@@ -116,7 +117,7 @@ export const ActantSuggester: React.FC<ActantSuggesterI> = ({
       }
       onCreate={(newCreated: {
         label: string;
-        category: "P" | "G" | "O" | "C" | "L" | "V" | "E";
+        category: CategoryActantType;
       }) => {
         handleCreate(newCreated);
       }}
