+++ packages/client/webpack.common.js
@@ -1,56 +1,60 @@
 const HtmlWebpackPlugin = require("html-webpack-plugin");
 const { CleanWebpackPlugin } = require("clean-webpack-plugin");
+const path = require("path");
 
 module.exports = {
-    entry: "./src/index.tsx",
-    module: {
-        rules: [
-            { test: /\.(js)$/, use: "babel-loader", exclude: /node_modules/ },
-            {
-                test: /\.css$/,
-                use: [
-                    "style-loader",
-                    {
-                        loader: "css-loader",
-                        options: {
-                            importLoaders: 1,
-                        },
-                    },
-                    "postcss-loader",
-                ],
+  entry: "./src/index.tsx",
+  module: {
+    rules: [
+      { test: /\.(js)$/, use: "babel-loader", exclude: /node_modules/ },
+      {
+        test: /\.css$/,
+        use: [
+          "style-loader",
+          {
+            loader: "css-loader",
+            options: {
+              importLoaders: 1,
             },
-            {
-                test: /\.tsx?$/,
-                use: "ts-loader?configFile=tsconfig.json",
-            },
-            {
-                test: /\.svg$/,
-                use: [
-                    {
-                        loader: "svg-url-loader",
-                        options: {
-                            limit: 10000,
-                        },
-                    },
-                ],
-            },
-            {
-                test: /\.(woff|woff2|ttf)$/,
-                use: {
-                    loader: "url-loader",
-                },
-            },
+          },
+          "postcss-loader",
         ],
-    },
-    resolve: {
-        extensions: [".tsx", ".ts", ".js"],
-        modules: ["node_modules", "src"],
-    },
-    mode: "development",
-    plugins: [
-        new CleanWebpackPlugin(),
-        new HtmlWebpackPlugin({
-            template: "./public/index.html",
-        }),
+      },
+      {
+        test: /\.tsx?$/,
+        use: "ts-loader?configFile=tsconfig.json",
+      },
+      {
+        test: /\.svg$/,
+        use: [
+          {
+            loader: "svg-url-loader",
+            options: {
+              limit: 10000,
+            },
+          },
+        ],
+      },
+      {
+        test: /\.(woff|woff2|ttf)$/,
+        use: {
+          loader: "url-loader",
+        },
+      },
     ],
+  },
+  resolve: {
+    extensions: [".tsx", ".ts", ".js"],
+    modules: ["node_modules", "src"],
+    alias: {
+      "@shared": path.resolve(__dirname, "../shared/"),
+    },
+  },
+  mode: "development",
+  plugins: [
+    new CleanWebpackPlugin(),
+    new HtmlWebpackPlugin({
+      template: "./public/index.html",
+    }),
+  ],
 };
