+++ packages/shared/enums/index.ts
@@ -1,0 +1,30 @@
+export enum ActantType {
+  Territory = "T",
+  Statement = "S",
+  Resource = "R",
+  Person = "P",
+  Group = "G",
+  Object = "O",
+  Concept = "C",
+  Location = "L",
+  Value = "V",
+  Event = "E",
+}
+
+export type CategoryActantType =
+  | ActantType.Person
+  | ActantType.Group
+  | ActantType.Object
+  | ActantType.Concept
+  | ActantType.Location
+  | ActantType.Value
+  | ActantType.Event;
+
+export type EntityActantType =
+  | ActantType.Person
+  | ActantType.Group
+  | ActantType.Object
+  | ActantType.Concept
+  | ActantType.Location
+  | ActantType.Value
+  | ActantType.Event;
