+++ packages/shared/types/actant.ts
@@ -1,15 +1,4 @@
-export enum ActantType {
-  Territory = "T",
-  Statement = "S",
-  Resource = "R",
-  Person = "P",
-  Group = "G",
-  Object = "O",
-  Concept = "C",
-  Location = "L",
-  Value = "V",
-  Event = "E",
-}
+import { ActantType } from "../enums";
 
 export interface IActant {
   id: string;
