+++ packages/shared/types/index.ts
@@ -1,6 +1,6 @@
 import { IAudit } from "./audit";
 import { IAction } from "./action";
-import { IActant, ActantType } from "./actant";
+import { IActant } from "./actant";
 import { IEntity } from "./entity";
 import { ILabel } from "./label";
 import { IProp } from "./prop";
@@ -27,7 +27,6 @@ import { IOption } from "./option";
 import { IResponseGeneric } from "./response-generic";
 
 export type {
-  ActantType,
   IAudit,
   IAction,
   IActant,
