+++ packages/shared/types/territory.ts
@@ -1,6 +1,6 @@
 import { IActant } from "./actant";
 import { languageDict } from "./../dictionaries";
-import { ActantType } from ".";
+import { ActantType } from "../enums";
 
 const languageValues = languageDict.map((i) => i.value);
 export interface ITerritory extends IActant {
